work on filter date flow logic in filterLandingPage.ts
