+++ holidaze-examproject/src/pages/LandingPage.tsx
@@ -1,3 +1,4 @@
+// LandingPage.tsx
 import { useEffect, useState, useCallback } from 'react';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid2';
@@ -10,6 +11,8 @@ import debounce from '../services/utilities/debounce.ts';
 import VenueData from '../services/interfaces/api/venueResponse.ts';
 import VenueQueryParams from '../services/interfaces/api/venueQueryParams.ts';
 import availableCountries from '../services/interfaces/api/filtering/availableCountries.ts';
+import { processVenueData } from '../services/filtering/filterLandingPage.ts';
+
 const apiKey = import.meta.env.VITE_NOROFF_API_KEY;
 
 export default function LandingPage() {
@@ -17,7 +20,6 @@ export default function LandingPage() {
   const [loading, setLoading] = useState<boolean>(false);
   const [error, setError] = useState<Error | null>(null);
   const [searchTerm, setSearchTerm] = useState<string>('');
-
   const headers = getValidatedHeader();
 
   const fetchVenueData = useCallback(
@@ -33,7 +35,7 @@ export default function LandingPage() {
         };
 
         const endpoint = venuesEndpoint(queryParams);
-        console.log(endpoint);
+        //console.log(endpoint);
 
         const response = await baseApiCall({
           url: endpoint,
@@ -71,6 +73,13 @@ export default function LandingPage() {
     }, 200),
     [headers],
   );
+
+  // Call the processVenueData function each time venueData is updated
+  useEffect(() => {
+    if (venueData.length > 0) {
+      processVenueData(venueData);
+    }
+  }, [venueData]);
 
   useEffect(() => {
     fetchVenueData(searchTerm);
