+++ holidaze-examproject/src/services/filtering/filterLandingPage.ts
@@ -1,0 +1,20 @@
+// filterLandingPage.ts
+import FilterParams from '../interfaces/api/filtering/uiFilterParams.ts';
+import VenueData from '../interfaces/api/venueResponse.ts';
+
+// Function to handle the filter logic
+export function applyFilters(filters: FilterParams) {
+  const {
+    availableChecked = false, // default to false if not provided
+    selectedCountries = [], // default to empty array if not provided
+    dates = { checkInDate: null, checkOutDate: null }, // default empty dates if not provided
+  } = filters || {};
+
+  console.log('Values to apply FILTER', availableChecked, selectedCountries, dates);
+
+}
+
+// Optional: Process venue data
+export function processVenueData(venueData: VenueData[]) {
+  console.log('Venue data to be FILTERED:', venueData);
+}
